refactor: move start task logic to taskReducer function
