+++ src/contexts/TaskContext/index.tsx
@@ -1,17 +1,18 @@
 import { createContext } from 'react';
 import type { TaskStateModel } from '../../models/TaskStateModel';
 import { initialTaskState } from './initialTaskState';
+import type { TaskActionModel } from './taskAction';
 
 // Type Props & initial Props for Context
 
 type TaskContextProps = {
   state: TaskStateModel;
-  setState: React.Dispatch<React.SetStateAction<TaskStateModel>>;
+  dispatch: React.Dispatch<TaskActionModel>;
 };
 
 const initialContextValue = {
   state: initialTaskState,
-  setState: () => {},
+  dispatch: () => {},
 };
 
 // Context creation
