+++ src/contexts/TaskContext/taskReducer.ts
@@ -1,4 +1,6 @@
 import type { TaskStateModel } from '../../models/TaskStateModel';
+import { formatSecondsToMinutes } from '../../utils/formatSecondsToMinutes';
+import { getNextCycle } from '../../utils/getNextCycle';
 import { TaskActionTypes, type TaskActionModel } from './taskAction';
 
 export function taskReducer(
@@ -7,11 +9,28 @@ export function taskReducer(
 ): TaskStateModel {
   switch (action.type) {
     case TaskActionTypes.START_TASK: {
-      return state;
+      const newTask = action.payload;
+      const secondsRemaining = newTask.duration * 60;
+      const nextCycle = getNextCycle(state.currentCycle);
+      const formattedSecondsRemaining = formatSecondsToMinutes(
+        state.secondsRemaining,
+      );
+
+      return {
+        ...state,
+        tasks: [...state.tasks, newTask],
+        secondsRemaining,
+        formattedSecondsRemaining, // ---- TODO ----
+        activeTask: newTask,
+        currentCycle: nextCycle,
+        config: { ...state.config },
+      };
     }
+
     case TaskActionTypes.INTERRUPT_TASK: {
       return state;
     }
+
     case TaskActionTypes.RESET_STATE: {
       return state;
     }
